git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1692593 13f79535-47bb-0310-9956-ffa450edef68

+++ src/scratchpad/src/org/apache/poi/hslf/model/textproperties/IndentProp.java
@@ -41,7 +41,7 @@ public class IndentProp  {
     public int getCharactersCovered() { return charactersCovered; }
     
     public int getIndentLevel() {
-        return (int)indentLevel;
+        return indentLevel;
     }
     
     /**
